Move `read_data` funcs into `read` submodule of `data`
